Moved android app module to android>mobile for better consistency with planned multi-form factor organization updated todo updated readme and modules dependencies graph
