+++ android/mobile/app/proguard-rules.pro
