+++ android/mobile/app/schemas/fr.shining_cat.simplehiit.data.local.database.SimpleHiitDatabase/1.json
