+++ buildSrc/src/main/kotlin/test_modules_gradle_config.gradle.kts
@@ -14,7 +14,7 @@ android {
         testInstrumentationRunner = "fr.shining_cat.simplehiit.testutils.HiltTestRunner"
     }
 
-    targetProjectPath = ":app"
+    targetProjectPath = ":android:mobile:app"
 
     buildTypes {
         getByName("debug") {
