+++ settings.gradle.kts
@@ -1,7 +1,11 @@
 rootProject.name = "SimpleHIIT"
-include(":app")
-include(":commonUtils")
-include(":testUtils")
+include(":android:mobile:app")
+include(":android:mobile:ui:common")
+include(":android:mobile:ui:home")
+include(":android:mobile:ui:session")
+include(":android:mobile:ui:settings")
+include(":android:mobile:ui:statistics")
+include(":commonResources")
 include(":domain:common")
 include(":domain:home")
 include(":domain:settings")
@@ -9,9 +13,5 @@ include(":domain:statistics")
 include(":domain:session")
 include(":data")
 include(":dataInstrumentedTests")
-include(":commonResources")
-include(":android:mobile:ui:common")
-include(":android:mobile:ui:home")
-include(":android:mobile:ui:session")
-include(":android:mobile:ui:settings")
-include(":android:mobile:ui:statistics")
+include(":testUtils")
+include(":commonUtils")
